Move auth and users controller from auth to a trait
